==> Nevada/WaterAllocation/2_NVwr_CreateWaDEInputFiles.ipynb <==
# Create WaDE Uploader csv Files for Water Right Data
- Last Updated: 06/23/2023
- Purpose: To create necessary processed WaDE Uploader csv files for water right information.

# Needed Libraries / Modules
import sys
import os

## Custom Libraries
sys.path.append("../../5_CustomFunctions/MappingFunctions")
import CreateWaterSourcesFile
import CreateSitesFile
import CreateAllocationsAmounts_factsFile
import RemoveUnusedRecordsFile
import CreatePODSiteToPOUSiteRelationshipsFile
import JoinOverlayToSiteFile
import AddCountyHUC8HUC12File

## Input Data (make changes here)
 - create and add variables specific to this proejct here

# ---- Inputs ----
varST = "NV" # source organization or state abbreviation
varUUIDType = "wr" # UUID data type abbreviation
mainInputFile = "Pwr_Main.zip" # use processed zip file

# ---- working directory ----
workingDirString = "G:/Shared drives/WaDE Data/WaDE Data Folder/Nevada/WaterAllocation"  # set working directory folder string here
os.chdir(workingDirString)
print(f'The working Directory is:', workingDirString)

## Create Water Right Input Files

# ---- Method (methods.csv) ----
# Simple info, create by hand and save in ProcessedInputData folder.

# ---- Variable (variables.csv) ----
# Simple info, create by hand and save in ProcessedInputData folder.

# ---- Organization (organizations.csv) ----
# Simple info, create by hand and save in ProcessedInputData folder.

# ---- Water Sources (watersources.csv) ----
print("Creating Water Sources input csv...")
print("############################################################################")
CreateWaterSourcesFile.CreateWaterSourcesInputFunction(workingDirString, varST, varUUIDType, mainInputFile)

# ---- Sites (sites.csv) ----
print("Creating Sites input csv...")
print("############################################################################")
CreateSitesFile.CreateSitesInputFunction(workingDirString, varST, varUUIDType, mainInputFile)

# ---- Allocations Amounts (waterallocations.csv) ----
print("Creating AllocationsAmounts_facts input csv...")
print("############################################################################")
CreateAllocationsAmounts_factsFile.CreateAllocationsAmounts_factsInputFunction(workingDirString, varST, varUUIDType, mainInputFile)

# ---- Remove unused records ----
print("Remove unused Water Sources and Sites records not found within AllocationsAmounts_facts input csv...")
print("############################################################################")
RemoveUnusedRecordsFile.RemoveUnusedAllocationsAmountRecordsFileFunction(workingDirString)

# ---- Create POD and POU joins (podsitetopousiterelationships.csv) ----
print("Creating Pod site -to- Pou site relationships input csv...")
print("############################################################################")
CreatePODSiteToPOUSiteRelationshipsFile.PODToPOUWaterRightRelationshipsFunction(workingDirString)

# ---- Join RegulatoryOverlayUUIDs for Sites (Sites.csv update) ----
print("Joining RegulatoryOverlayUUIDs to Sites input csv if exists...")
print("############################################################################")
JoinOverlayToSiteFile.JoinOverlayToSiteFunction(workingDirString)

# ---- Add missing County, HUC8, HUC12 information for Sites (Sites.csv update) ----
print("Add misinsg County, HUC8, HUC12 information if mising...")
print("############################################################################")
AddCountyHUC8HUC12File.AddCountyHUC8HUC12Function(workingDirString)
