==> Nevada/WaterAllocation/4_WaterSources.py <==
#Date Created: 06/12/2020
#Purpose: To extract TX water source use information and population dataframe for WaDE_QA 2.0.
#Notes: 1) For 'WaterSourceTypeCV', easier to label everything that is not a surface water first.

# Note:  Currently using dummy row until more data made available from TX

# Needed Libraries
############################################################################
import pandas as pd
import numpy as np
import os


# Inputs
############################################################################
print("Reading input csv...")
workingDir="/Users/augustus/Desktop/WSWC/WaDE/MappingStatesDataToWaDE2.0/Nevada/WaterAllocation/"
os.chdir(workingDir)
#fileInput = "RawinputData/P_IdahoMaster.csv"
#df = pd.read_csv(fileInput)

#WaDE columns
columnslist = [
    "WaterSourceUUID",
    "Geometry",
    "GNISFeatureNameCV",
    "WaterQualityIndicatorCV",
    "WaterSourceName",
    "WaterSourceNativeID",
    "WaterSourceTypeCV"]


# Custom Site Functions
############################################################################

# For creating WaterSourceName
def assignWaterSourceName(colrowValue):
    if colrowValue == '' or pd.isnull(colrowValue):
        outList = "Unknown"
    else:
        outList = colrowValue
    return outList

# For creating WaterSourceTypeCV
UnknownWSCVDict = {
"Ground Water" : "Ground Water",
"Spring" : "groundwater/spring",
"Waste Water" : "Reuse",
"Wasteway" : "Reuse",
"Wastewater" : "Reuse",
"Drain" : "Drain",
"Reservoir" : "reservoir"
}
def assignWaterSourceTypeCV(colrowValue):
    if colrowValue == "" or pd.isnull(colrowValue):
        outList = ""
    else:
        String1 = colrowValue.strip()  # remove whitespace chars
        try:
            outList = UnknownWSCVDict[String1]
        except:
            outList = "Surface Water"
    return outList

# For creating WaDESiteUUID
def assignWaterSourceUUID(colrowValue):
    string1 = str(colrowValue)
    outstring = "ID_" + string1
    return outstring


# Creating output dataframe (outdf)
############################################################################
print("Populating dataframe...")
outdf=pd.DataFrame(columns=columnslist)
outdf = outdf.append({"WaterSourceUUID":'NV_1',
    "Geometry":'',
    "GNISFeatureNameCV":'',
    "WaterQualityIndicatorCV":'Unspecified',
    "WaterSourceName":'',
    "WaterSourceNativeID":'',
    "WaterSourceTypeCV":''}, ignore_index=True)
"""
outdf = pd.DataFrame(index=df.index, columns=columnslist)  # The output dataframe for CSV.


print("Geometry")  # Hardcoded
outdf.Geometry = ""

print("GNISFeatureNameCV")  # Hardcoded
outdf.Geometry = ""

print("WaterQualityIndicatorCV")  # Hardcoded
outdf.WaterQualityIndicatorCV = "Unspecified"

print("WaterSourceName")
#outdf['WaterSourceName'] = df.apply(lambda row: assignWaterSourceName(row['TributaryOf']), axis=1)
outdf['WaterSourceName'] = ""

print("WaterSourceNativeID")  # has to be one of the last, need length of created outdf
outdf['WaterSourceNativeID'] = df['WaterDistrictNumber'].str.upper()


print("WaterSourceTypeCV")
outdf['WaterSourceTypeCV'] = df.apply(lambda row: assignWaterSourceTypeCV(row['Source']), axis=1)


##############################
# Dropping duplicate
print("Dropping duplicates")
outdf = outdf.drop_duplicates(subset=['WaterSourceName', 'WaterSourceTypeCV']).reset_index(drop=True)
##############################

print("WaterSourceUUID") #native source identifier and the organization univeral id. has to be one of the last, need WaterSourceNativeID to create
df["Count"] = range(1, len(df.index) + 1)
outdf['WaterSourceUUID'] = df.apply(lambda row: assignWaterSourceUUID(row['Count']), axis=1)


# Check required fields are not null
############################################################################
#Check all 'required' (not NA) columns have value (not empty). Replace blank strings by NaN, if there are any
print("Checking required is not null...")
outdf = outdf.replace('', np.nan)  # Replace blank strings by NaN, if there are any
outdf100_nullMand = outdf.loc[(outdf["WaterSourceUUID"].isnull()) | (outdf["WaterSourceUUID"] == '') |
                              (outdf["WaterSourceTypeCV"].isnull()) | (outdf["WaterSourceTypeCV"] == '') |
                              (outdf["WaterQualityIndicatorCV"].isnull()) | (outdf["WaterQualityIndicatorCV"] == '')]


# Checking Dataframe data types
############################################################################
print(outdf.dtypes)


# Export to new csv
############################################################################


#Report missing values if need be to seperate csv
if(len(outdf100_nullMand.index) > 0):
    outdf100_nullMand.to_csv('watersources_mandatoryFieldMissing.csv')  # index=False,
"""
print("Exporting dataframe to csv...")
outdf.to_csv('ProcessedInputData/watersources.csv', index=False)

print("Done.")
